+++ qoqo_qiskit_devices/src/devices/ibm_lima.rs
@@ -35,6 +35,9 @@ impl IBMLimaDeviceWrapper {
     /// Create a new IBMLimaDevice instance.
     #[new]
     pub fn new() -> Self {
+        Python::with_gil(|py| {
+            py.run("import warnings; warnings.warn(\"Device ibmq_lima has been retired. Setting noise information is not possible.\", category=DeprecationWarning, stacklevel=2)", None, None).unwrap();
+        });
         Self {
             internal: IBMLimaDevice::new(),
         }