+++ qoqo_qiskit_devices/src/devices/ibm_manila.rs
@@ -35,6 +35,9 @@ impl IBMManilaDeviceWrapper {
     /// Create a new IBMManilaDevice instance.
     #[new]
     pub fn new() -> Self {
+        Python::with_gil(|py| {
+            py.run("import warnings; warnings.warn(\"Device ibmq_manila has been retired. Setting noise information is not possible.\", category=DeprecationWarning, stacklevel=2)", None, None).unwrap();
+        });
         Self {
             internal: IBMManilaDevice::new(),
         }