+++ qoqo_qiskit_devices/src/devices/ibm_quito.rs
@@ -35,6 +35,9 @@ impl IBMQuitoDeviceWrapper {
     /// Create a new IBMQuitoDevice instance.
     #[new]
     pub fn new() -> Self {
+        Python::with_gil(|py| {
+            py.run("import warnings; warnings.warn(\"Device ibmq_quito has been retired. Setting noise information is not possible.\", category=DeprecationWarning, stacklevel=2)", None, None).unwrap();
+        });
         Self {
             internal: IBMQuitoDevice::new(),
         }