+++ roqoqo_qiskit_devices/src/devices/ibm_belem.rs
@@ -39,6 +39,7 @@ impl IBMBelemDevice {
     ///
     /// An initiated IBMBelemDevice with single and two-qubit gates and decoherence rates set to zero.
     ///
+    #[deprecated(since = "0.2.0", note = "Device ibmq_belem has been retired.")]
     pub fn new() -> Self {
         let mut device = Self {
             number_qubits: 5,
@@ -273,6 +274,7 @@ impl QoqoDevice for IBMBelemDevice {
             "PauliX".to_string(),
             "RotateZ".to_string(),
             "SqrtPauliX".to_string(),
+            "Identity".to_string(),
         ]
     }