+++ roqoqo_qiskit_devices/tests/integration/main.rs
@@ -10,6 +10,7 @@
 // express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
 
+#[allow(deprecated)]
 #[cfg(test)]
 mod device;
 pub use device::*;